==> 02_dataclean.Rmd <==
---
title: "02_Data cleaning_First(Errors and Time)"
output: html_document
---
 

# Data cleaning before id_classification and model  
# 데이터 클리닝 


```{r}
df<-read.csv("wowdata06.csv")

library(dplyr)

#특정 열 기준 불필요 데이터 제거 
df <- df[complete.cases(df[,6]),]

#dummy 제거 
df <- df[,-1] 

#이름 변경 
names(df) <- c("Query_time","SequenceNum","AvatarId","Guild" , "Level", "Race", "Class", "Zone"  )

#빈 칸 제거 
df$Query_time <-str_trim(df$Query_time)
df$SequenceNum <- str_trim(df$SequenceNum)
df$AvatarId <- str_trim(df$AvatarId)
df$Guild <- str_trim(df$Guild)
df$Level <- str_trim(df$Level)
df$Race <- str_trim(df$Race)
df$Class <- str_trim(df$Class)
df$Zone <- str_trim(df$Zone)
```


#for checking whether whole year data is included  
#일 년의 데이터가 들어온 것을 확인하기 


```{r}
#시간 데이터 분리 
df$month <- substr(df$Query_time,1,2)
df$day <- substr(df$Query_time,4,5)
df$time <- substr(df$Query_time,10,17)

#12개월 데이터가 다 들어있는지 확인 
month <-substr((df[,1]), 2,3) 
unique(month)
```



#Checking and eliminating for value error     
#에러값 확인 및 제거  
```{r}
#Race 이상치 확인 
unique(df$Race)
df <- df %>%
  filter(Race %in% c("Orc", "Tauren", "Troll", "Undead", "Blood Elf" ) )
unique(df$Race)

#Class 이상치 확인 
unique(df$Class)
df <- df %>%
  filter(Class!= "2400") 
unique(df$Class)

#쿼리타임 순 배열 
df <- df %>%
  arrange(Query_time)
```

```{r}
write.csv(df,'000_wowdata_0708.csv' )
```




# Exploratory Search and detecting real starting users
# EDA 및 실제 시작유저 탐색, 초기 유저 정의 
 
- Discriminating sub characters for item trading/storing from real users(including second character to play) 
-Finding out proportion of those who left in the middle of game by graph  

- WOW 내 창고 캐릭터 제외, 실제 유저(육성용 부캐 포함) 가르기
-그래프 확인 및 중간 이탈자 확인 


```{r}
dat <- read.csv('000_wowdata_0708.csv')
```

```{r}
data  <- dat[,5:9]
data1 <- data[,-2]
data1 <- unique(data1)

#레벨 분리해내기 
data1 <-data1 %>% 
  group_by(AvatarId,Class)%>%
  mutate(minLev = min(Level))%>%
  mutate(maxLev = max(Level))%>%
  mutate(levup = max(Level) - min(Level))

data1_id <- data1[,-2]
data1_id <- unique(data1_id)

data_noleveup <- data1_id%>%
  filter(levup<=0)
data_levup <- data1_id %>%
  filter(levup!=0)


#initial status  시작 상태 정의하기 
# 레벨 3 이하 레벨 업 5 이상 ~ 실제 사용 유저 
data_ini <- data_levup %>%
  filter(minLev<=3 & levup>=5)
```


```{r}
library(ggplot2)
inilevup <- ggplot(data= data_ini) + geom_density(aes(x=maxLev),fill="#69b3a2", color="#e9ecef", alpha=0.8)+
 labs(title="Final levelup point of new users(Lev 1~3)", x = "min Level", y = "density")  
inilevup
inilevup2 <- ggplot(data= data_ini) + geom_bar(aes(x=maxLev),fill="#6600CC", color="#9933FF", alpha=0.8)+
 labs(title="The number of new users reaching levelup point", x = "min Level", y = "number of users")  
inilevup2
```

#EDA  
From graph we could find out that initial users played until level 7~10 and decided leave or not, so Level 6~10 become period for user type observation. 

그래프 상 레벨 7~10까지 플레이 후 게임 이탈 결정이 대거 이뤄지므로, 레벨 6~10 선이 유저 타입의 관찰기간이 됨. 




#Data cleaining for QueryTime 
#시간 데이터 뽑아내기 
```{r}
id <- data_ini[,1]
id <- as.data.frame(id)
```

```{r}
library(stringr)
dat1  <- dat[,3:10]
dat1  <- dat1[,-2]

id2 <- left_join(id, dat1, by = 'AvatarId')
id2$QueryTime <- str_trim(id2$QueryTime)
id2$year <- substr(id2$QueryTime,7,8)
id2$month <- substr(id2$QueryTime,1,2)
id2$day <- substr(id2$QueryTime,4,5)
id2$time <- substr(id2$QueryTime,10,17)
id2$Qtime <-paste(id2$year, id2$month, id2$day, sep= "-")
id2$Qtime <-paste0(20, id2$Qtime)
id2$QT <- paste(id2$Qtime, id2$time, sep=" ")

id2$QT2 <- as.POSIXct(id2$QT, format='%Y-%m-%d %H:%M:%S')

id2 <- id2 %>%
  group_by(AvatarId, QT)%>% 
  arrange(AvatarId, QT)

data_time<-id2[,-2]  
data_time<-data_time[,-7]  
data_time<-data_time[,-7]
data_time<-data_time[,-7]
data_time<-data_time[,-7]
data_time<-data_time[,-7]
data_time<-data_time[,-7]
names(data_time)[7] <- c("QT")
data_time <- unique(data_time)
```

```{r}
library(lubridate)
data_time <- data_time %>%
  group_by(AvatarId,Class, QT) %>%
  arrange(AvatarId,Class, QT)

data_time$lagtime <- lag(data_time$QT, n=1L)
data_time$timeraw <- data_time$QT - data_time$lagtime

data_time$y <- paste0(data_time$AvatarId, data_time$Race, data_time$Class)
data_time$ly <- lag(data_time$y, n=1L)
data_time$timeraw <- ifelse(data_time$y!=data_time$ly, 0, data_time$timeraw)

data_time$minraw <-seconds_to_period(data_time$timeraw) 
data_time <- data_time[,-c(11,12,13)]
```


#Combinging Region 
#지역결합
* You can find zones.csv from kaggle. 

```{r}
zone <-read.csv("zones.csv")
zone1 <-zone[,c("Zone_Name","Type","Controlled")]

data_time$Zone=str_trim(data_time$Zone)
data_t<-left_join(data_time,zone1,by=c('Zone'='Zone_Name'))
```

```{r}
write.csv(data_t, "05time_zone.csv")
write.csv(data_time, "06time.csv") 
```
