==> 03_eda_classification.Rmd <==
---
title: "03_eda and kmeans classification"
author: "Inwon"
date: '2022 3 21' 
---


# Exploratory Search and detecting real starting users
# EDA 및 실제 시작유저 탐색, 초기 유저 정의 
 
- Discriminating sub characters for item trading/storing from real users(including second character to play) 
-Finding out proportion of those who left in the middle of game by graph  

- WOW 내 창고 캐릭터 제외, 실제 유저(육성용 부캐 포함) 가르기
-그래프 확인 및 중간 이탈자 확인 


```{r}
dat <- read.csv('000_wowdata_0708.csv')
dat <-dat[,-1]
```

#Data cleaining for QueryTime 
#시간 데이터 뽑아내기 

```{r}
library(stringr)
library(dplyr)

dat$id <- paste0(dat$AvatarId,dat$Race,dat$Class)

dat$QueryTime <- str_trim(dat$QueryTime)
dat$year <- substr(dat$QueryTime,7,8)
dat$month <- substr(dat$QueryTime,1,2)
dat$day <- substr(dat$QueryTime,4,5)
dat$time <- substr(dat$QueryTime,10,17)
dat$Qtime <-paste(dat$year, dat$month, dat$day, sep= "-")
dat$Qtime <-paste0(20, dat$Qtime)
dat$QT <- paste(dat$Qtime, dat$time, sep=" ")

dat$QT2 <- as.POSIXct(dat$QT, format='%Y-%m-%d %H:%M:%S')

dat <- dat[,-c(1,2,3,4,11,12,13,14,15,16)]  
names(dat[7])= c("QT")
```

#eliminateing error value 
```{r}
#이상치 날리기 
unique(dat$Race)
```
```{r}
dat <- dat %>%
  filter(Race %in% c(" Orc"," Tauren"," Troll" , " Undead", " Blood Elf"))
```
```{r}
unique(dat$Class) 
```
```{r}
dat <- dat%>%
  filter(Class != "2400")
```

```{r}
#해외 조닝 날리기 
dat$id = str_trim(dat$id)
dat$Race = str_trim(dat$Race)
dat$Class = str_trim(dat$Class)
dat$Zone = str_trim(dat$Zone)
dat$Guild = str_trim(dat$Guild)
dat$Level = str_trim(dat$Level)

zone <- read.csv("zones.csv")
zone <- zone[,c(1,6,8)] 
names(zone[3])=c("Zone")

dat <-inner_join(dat,zone, by=c("Zone"="Zone_Name"))
```

```{r}
#길드 결측치 제거하기 
dat <-dat %>% 
  filter(!is.na(Level))
dat$Guild[is.na(dat$Guild)]<-0  

dat <- na.omit(dat)
```

```{r}
#창고캐 (쓰지 않는 캐릭터) 정리하기 

dat$Level <- as.numeric(dat$Level)

dat <- dat %>% 
  group_by(id)%>% 
  mutate(minLev= min(Level) , maxLev = max(Level), Levup =max(Level)-min(Level))

dat <- dat %>%
  filter(Levup >=3 | minLev>3)
```

```{r}
unique(dat$Type) 
#Zone : Dungeon 던전 ; Battleground 팀 전장 ; Arena 개인 전장 ; Sea 바다 관련 퀘스트; Event  ; Transit 
```






#주간 데이터 생성하기
 - 절대주간(week). 상대주간(t)
 - 해당 주 길드 인원 수
 - 길드에 만렙 유무 

```{r}
#절대주간, 상대주간 생성하기 
library(lubridate)
library(dplyr)

#절대주간 week 
dat$week <- week(dat$QT2)
dat$tWeek <- ifelse(dat$week-12>=0, dat$week-12 , dat$week+41) 
dat$tWeek <- ifelse(dat$tWeek==0,53,dat$tWeek) 

dat <- dat%>% 
  filter(substr(QT2,1,10) != "2008-03-30")
dat <- dat%>% 
  filter( substr(QT2,1,10) != "2008-03-31")
dat <- dat[,-14]


#상대주간 t 
dat <-dat%>%
  group_by(id)%>%
  mutate(mint = min(week))

dat <-dat%>%
  group_by(id)%>%
  mutate(t = ifelse(mint==1, week, week-mint+1))
dat<-dat[,-14]

```

```{r}
#절대주간 기준 길드 인원 수 추출하기 
dat_id<- dat[,c('week','Guild', 'id' ,"maxLev")]
dat_id <- unique(dat_id)

#주별 길드별 인원 추출, 최고레벨 추출 
dat_id <- dat_id %>%
  group_by(week,Guild)%>%
  mutate(gm = n(),maxLg =max(maxLev))

unique(dat_id$maxLg) #만렙만 있는 길드만 있는 건 아님 

#길드 만렙 유무 
#만렙자 보유 길드는 =1, 아니면 = 0 
dat_id$manlv <- ifelse(dat_id$maxLg>=60,1,0) 

dat_id <- dat_id[,c(1,2,3,5,7)]
write.csv(dat_id, "30_dat_id.csv")
```



# 분석 대상 데이터 추출하기 

```{r}
#초기모델에서 레벨업 15 이상 한 사람을 추출하기 
dat2 <- dat%>%
  filter(Levup>15& minLev<=4)
```

```{r}
#4163명 
id <- unique(dat2$id)
write.csv(dat2, "31_targetusers.csv")
write.csv(dat, "29_rawdata.csv")
```



#여기서 시작 
```{r}
dat2 <-read.csv("31_targetusers.csv")
dat2<-dat2[,-1]
library(dplyr)
```


#주간 데이터 생성2 
 - 해당 주 전체 플레이타임 
 - 해당 주 로그기록 : Arena/Battleground, Dungeon/Sea(업적) 
```{r}
dat2 <- dat2 %>%
  group_by(id, t, QT2)%>% 
  arrange(id, t, QT2)
```


```{r}
#로그별 플레이 시간 추출하기 
library(lubridate)
dat2$QT2 <- as.POSIXct(dat2$QT2, format='%Y-%m-%d %H:%M:%S')
dat2$lagtime <- lag(dat2$QT2, n=1L)
dat2$timeraw <- dat2$QT2 - dat2$lagtime

dat2$ly <- lag(dat2$id, n=1L)
dat2$timeraw <- ifelse(dat2$id!=dat2$ly, 0, dat2$timeraw)

#dat2$minraw <-seconds_to_period(dat2$timeraw) 
dat2$minraw <- round((dat2$timeraw)/60) 

dat2 <- dat2[,-c(15:17)]
```

```{r}
#쉬는 시간 날리기 
dat2<-dat2%>%
  filter(minraw<=30 & minraw>0)
```
```{r}
#지역타입 추출하기 
arena= c("Arena") #팀전 
pvp = c("Battleground","Arena" )
<<<<<<< HEAD
dungeon = c("Dungeon")
=======

dungeon = c("Dungeon")

>>>>>>> 3735914ce0abdacf5aba29e312ace07e50d59376
quest = c("Dungeon","Sea", "Event") #싸움 퀘스트 수행지 
city = c("City") #경매장, 퀘스트  
others = c("Zone","Transit")
```

```{r}
#전체 플레이타임 및 타입별 플레이타임 추출하기 
dat3 <- dat2 %>% 
  group_by(id, t, Level, Type)%>%
  mutate(aretime = ifelse(Type %in% arena, sum(minraw), 0), 
         pvptime = ifelse(Type %in% pvp, sum(minraw), 0),
<<<<<<< HEAD

         duntime = ifelse(Type %in% dungeon, sum(minraw),0),

=======
         duntime = ifelse(Type %in% dungeon, sum(minraw),0),
>>>>>>> 3735914ce0abdacf5aba29e312ace07e50d59376
         quetime = ifelse(Type %in% quest, sum(minraw),0), 
         cittime = ifelse(Type %in% city, sum(minraw),0),
         othtime = ifelse(Type %in% others, sum(minraw),0))

<<<<<<< HEAD
data <- dat3[,c(14,13,6,2,1,16,17,18,19,20,21)]
data <- unique(data)

=======

data <- dat3[,c(14,13,6,2,1,16,17,18,19,20,21)]
data <- unique(data)


>>>>>>> 3735914ce0abdacf5aba29e312ace07e50d59376
data <- dat3[,c(14,13,6,2,1,16,17,18,19,20)]
data <- unique(data)

```

```{r}
#플레이타임들 한 행에 합치기 
data <- data%>%
  group_by(id,t,Level)%>%
  mutate(arena = sum(aretime),
         pvp= sum(pvptime), 
<<<<<<< HEAD
         dungeon = sum(duntime),
=======

         dungeon = sum(duntime),

>>>>>>> 3735914ce0abdacf5aba29e312ace07e50d59376
         quest = sum(quetime), 
         city = sum(cittime),
         other = sum(othtime), 
         tottime = sum(pvptime)+sum(quetime)+sum(cittime)+sum(othtime))

#주간 길드 여부 

data <- data[,-c(6:11)]  
<<<<<<< HEAD
data <- data[,-c(6:10)]  
=======

data <- data[,-c(6:10)]  

>>>>>>> 3735914ce0abdacf5aba29e312ace07e50d59376
data <- data %>% 
  group_by(id,t,Level)%>%
  mutate(guild = max(Guild))

#레벨별 시간 매기기 
data <- unique(data)
```


```{r}
##신규 이전 기존캐릭터 날리기 
data <- data %>%
  group_by(id)%>%
  mutate(lead = lead(Level,1), 
         fi = ifelse(Level >= lead(Level, 1), "f", "o"))

#배열 중 0이 아닌 숫자가 나오면 그 이후 남은 길이만큼 반복하는 함수   
dup <- function(x) {
  x <- as.vector(x)
  lead <- lead(x,1)
  list = list()
  len = length(x)[1]
  loc = which(x>0)[1]
  k = len-loc+1
  a = rep(0,loc-1)
  b = rep(x[loc], k)
  list = append(a,b)
  list = as.numeric(list)
  return(list)
}
#함수적용 - 최소 레벨이 있는 주 이전은 0, 이후는 최소 주수 반복 
data <- data %>%
  group_by(id)%>%
  mutate(minl = ifelse(Level == min(Level), week, 0))
data <- data%>% 
  group_by(id)%>%
  mutate(some = dup(minl))

data <- data%>%
  filter(some!=0)%>%
  group_by(id,t,Level)
```


```{r}
# 같은 주내 길드 중복 날리기 - 큰 숫자 순서 
data2 <- data[,-5]
data2 <- unique(data2)


#상대주간 t 재생성 
data2 <-data2%>%
  group_by(id)%>%
  mutate(mint = min(week))

data2 <-data2%>%
  group_by(id)%>%
  mutate(t = ifelse(mint==1, week, week-mint+1))

```
 
 
```{r}

dat <- data2[,-c(13:17)]
write.csv(dat, "32_refineddata_withd.csv")
<<<<<<< HEAD
dat <- data2[,-c(12:16)]
write.csv(dat, "32_refineddata.csv")
=======

dat <- data2[,-c(12:16)]
write.csv(dat, "32_refineddata.csv")

>>>>>>> 3735914ce0abdacf5aba29e312ace07e50d59376
```
 
 
 
 
 
# 레벨별 operational efficiency로 그룹 kmeans 분류하기 
```{r}  
#
datbyl <- data2 %>%
  group_by(id,Level)%>% 
  mutate(playtime = sum(tottime),
         arenatime = sum(arena),
         pvptime=sum(pvp))

datbyl <- datbyl[,c(1,3,4,17,18,19)]
datbyl <- unique(datbyl)
datbyl <- datbyl%>%
  filter(Level!=70)
```

```{r}
#plot(x=datbyl$Level, y=datbyl$playtime)

#레벨 csv 소환 
Levs <- read.csv("Levels.csv")
Levs <- Levs[,1:2] 
names(Levs) <- c("Level", "Xp")
datbyl <- left_join(datbyl , Levs, by=("Level"))
datbyl$Xp <- as.numeric(datbyl$Xp)

#operational efficiency by individual 
datunder60 <- datbyl%>%
  filter(Level<=60)%>%
  mutate(oe_it = Xp/playtime)   #oe :해당 레벨에서의 본인의 효용비 

#observation 수가 10 미만인 유저 제외 
datunder60 <-datunder60 %>%
  group_by(id)%>%
  mutate(n=n())
datunder60<-datunder60 %>%
  filter(n>=10)

#operational efficiency - mean, var, square term 
datunder60 <- datunder60%>%
  group_by(Level)%>%
  mutate(oemean = mean(oe_it), 
         oevar = sqrt(oe_it)^2, 
         oesquare_it= sqrt((oe_it-mean(oe_it))^2)) #oe_square_i:해당 레벨 평균과 차이 

#그때까지의 oe 평균내는 함수 만들기 

cummean <- function(x){
  len <- length(x)
  i = len
  vector = c()
  for (i in 1:len){
    xi <- x[1:i] 
    cumsum <- cumsum(xi)
    cs <- tail(cumsum,1)
    mean <- cs/i
    vector = append(vector,mean)
  }
  return(vector)
}

cumvar <- function(x){
  len <- length(x)
  i = len
  vector = c()
  for (i in 1:len){
    xi <- x[1:i] 
    cumsum <- cumsum(xi)
    cs <- tail(cumsum,1)
    mean <- cs/i
    var = sqrt((x[i]-mean)^2)
    vector = append(vector,var)
  }
  return(vector)
}

#operational efficiency - cummean 
datunder60 <- datunder60%>%
  arrange(id, Level)%>%
  group_by(id)%>% 
  mutate(oecummean_it = cummean(oe_it))#i의 t레벨까지의 누적평균 oe(history 인식)

datunder60 <- datunder60%>%
  group_by(id)%>% 
  mutate(oecumvar_it = cumvar(oecummean_it))#i의 t레벨까지의 누적분산 oe(history 인식)

```



# 1) 20 이하의 정보로 분류하기 
```{r}
#레벨 20이하, 관측치 15개 이상 고르기 
datunder20 <- datunder60 %>%
  filter(Level<=20)%>%
  group_by(id)%>%
  mutate(n=n())

datunder20 <- datunder20%>%
  filter(n>10)

id20 <- unique(datunder20$id)

#매트릭스 생성하기 
id <- rep(id20, each=20)
Level <- rep(1:20, length(id20))
uu20 <- cbind(id,Level) 
uu20 <- as.data.frame(uu20)

datunder20$Level <- as.numeric(datunder20$Level)
uu20$Level <- as.numeric(uu20$Level)

uu20 <- left_join(uu20, datunder20, by= c("id", "Level"))

#uu20$leave <- ifelse(is.na(uu20$oe_it),0,1)
uu20na <- na.omit(uu20)
uu20na$lag <- lag(uu20na$Level,1) 
uu20na$lag[is.na(uu20na$lag)]<-0
uu20na$leave<- ifelse(uu20na$Level==uu20na$lag+1,1,0)
```


# 2) 비지도학습 - 레벨 20이하 (초반)의 정보로 kmeans 하기.
 - IV: t=1~4까지의 개인의 operational efficiency 평균(oe_it), operational efficiency 편차의 평균(oesqaure_i),개인의 누적된 oe의 변동 중 최신값 
 - 선행연구에 따르면 experiencer일수록 초반에 allow marginal costs하다. 

```{r}
#스케일링  
uu20na2 <- uu20na %>%
  group_by(id)%>%
  mutate(oe_i = mean(oe_its), #개인의 operational efficiency 평균(oe_it)
        oesquare_i = mean(oesquare_its), # operational efficiency 편차의 평균(oesqaure_i)
        oecumvar_i = tail(oecumvar_its,1)) #개인의 누적된 oe 결과 중 최신
uu20na2 <- uu20na2[,c(1,20,21,22)]
uu20na2 <- unique(uu20na2)
uu20na2_scaled <- scale(uu20na2[2:4])
uu20na2_scaled <- as.data.frame(uu20na2_scaled)
uu20na2 <- cbind(uu20na2, uu20na2_scaled)
names(uu20na2) <- c("id", "oe_i", "oesquare_i","oecumvar_i", "oe_iScaled", "oesquare_iScaled", "oecumvar_iScaled" )
summary(uu20na2)
```
# 3) 최적 클러스터의 수 확인 = 2개 
```{r}
library(NbClust)
nc <- NbClust(uu20na2[,5:7], min.nc=2, max.nc=6, method = 'kmeans')
```
#4) kmeans 
```{r}
kmeans.uu20 <- kmeans(uu20na2[,5:7],center=2)
kmeans.uu20

# Cluster means:
#   oe_iScaled oesquare_iScaled oecumvar_iScaled
# 1 -0.3690279       -0.3668904       -0.3736622
# 2  1.5527695        1.5437757        1.5722696 
#effciency가 높은 2번이  achiever 

uu20na2$leave <- ifelse(kmeans.uu20$cluster==1,1,0)#1:experiencer 0: achiever

#경향성에 따른 최종 분류 
uu20na2 %>%
  group_by(leave)%>%
  summarise(n=n())

#g   n 
#0	2613 : Experiencer ! 수정 ! 
#1	621 : Achiever! 
```

```{r}
plot(uu20na2[c("oe_iScaled","oesquare_iScaled")], col= kmeans.uu20$cluster)
```
#5)이상치 제거하기 
```{r}
boxplot(oe_iScaled+oesquare_iScaled+oecumvar_iScaled~leave, data =uu20na2)

oeUC = quantile(uu20na2$oe_iScaled, probs=c(0.75), na.rm=TRUE)+1.5*(quantile(uu20na2$oe_iScaled, probs=c(0.75), na.rm=TRUE)-quantile(uu20na2$oe_iScaled, probs=c(0.25), na.rm=TRUE))

oesUC = quantile(uu20na2$oesquare_iScaled, probs=c(0.75), na.rm=TRUE)+1.5*(quantile(uu20na2$oesquare_iScaled, probs=c(0.75), na.rm=TRUE)-quantile(uu20na2$oesquare_iScaled, probs=c(0.25), na.rm=TRUE))
                                                                                    
oecUC = quantile(uu20na2$oecumvar_iScaled, probs=c(0.75), na.rm=TRUE)+1.5*(quantile(uu20na2$oecumvar_iScaled, probs=c(0.75), na.rm=TRUE)-quantile(uu20na2$oecumvar_iScaled, probs=c(0.25), na.rm=TRUE))

uu20uc.refined <- uu20uc%>%
  filter(oe_iScaled<2.35 & oesquare_iScaled < 1.86)


boxplot(oe_iScaled+oesquare_iScaled+oecumvar_iScaled~leave, data =uu20uc.refined)

```

```{r}
uu20uc.refined %>%
  group_by(leave)%>%
  summarise(n=n())
```


```{r}
write.csv(uu20na2, "33_kmeans20classification.csv")
```

```{r}
#위에서 잘못 분류한 것 고치기 
# datbyl <- read.csv("33_kmeans20classification.csv")
# datbyl$class = ifelse(datbyl$leave ==0, "E", "A")
# datbyl$leave = ifelse(datbyl$class =="E", 1,0)
# write.csv(datbyl, "33_kmeans20classification.csv")
```





 

#t기의 정보 활용하기.
 - IV: t=1~4까지의 주별 개인의 operational efficiency 평균(oe_it), operational efficiency 편차의 평균(oesqaure_it), operational efficiency의 분산(oevar_it), 레벨변화(levelup_it) 
 - DV : t기에 접속/미접속 여부 
 
여기서 시작 
```{r}
data2 <- read.csv("32_refineddata.csv")
data2 <- unique(data2[,-1])
```
 
 #1) 변수 만들기 
```{r}
#데이터 재소환 
datbytl <- data2 %>%
    filter(Level!=60 & t <=4) 

datbytl <- datbytl[,-c(2,11,12,13,14,15,16)]
datbytl <- unique(datbytl) #길드 중복 없애기 
```

```{r}
#레벨 csv 소환 
Levs <- read.csv("Levels.csv")
Levs <- Levs[,1:2] 
names(Levs) <- c("Level", "Xp")
datbytl <- left_join(datbytl , Levs, by=("Level"))
datbytl$Xp <- as.numeric(datbytl$Xp)

#operational efficiency by individual 
datbytl <- datbytl%>%
  filter(Level<=60)%>%
  mutate(oe_il = Xp/tottime)   #oe :해당 레벨에서의 본인의 효용비 

#observation 수가 10 미만인 유저 제외 
datbytl <-datbytl %>%
  group_by(id)%>%
  mutate(n=n())
datbytl<-datbytl %>%
  filter(n>=10)

#operational efficiency - mean, var, square term 
datbytl <- datbytl%>%
  group_by(Level)%>%
  mutate(oemean = mean(oe_il), 
         oesd = sd(oe_il), 
         oesquare_il= sqrt((oe_il-mean(oe_il))^2)) #oe_square_il:해당 레벨 평균과 거리  

#operational efficiency - i의 t주 내 변수로 정리 
datbytl <- datbytl%>%
  arrange(id, t, Level)%>%
  group_by(id,t)%>% 
  mutate(oe_it = mean(oe_il), #i의 t주 동안 레벨 평균 oe 
         oesquare_it =mean(oesquare_il),#i의 t주 내 operational efficiency 편차의 평균(oesqaure_it)
         oevar_it = var(oe_il, na.rm = TRUE),  #operational efficiency의 분산(oevar_it)
         levup_it = max(Level)- min(Level)) #i의 t주 내 레벨업 

datbyt <- datbytl[,c(1,2,16,17,18,19)]
datbyt <- unique(datbyt)

#t=1 4주 모두 접속 t=0 4주 내 결석 존재 
datbyt <- datbyt %>%
  group_by(id)%>%
  mutate(n = ifelse(n()==4, 1,0))

#na 그룹별 평균값으로 대체 
datbyt <- datbyt %>%
  group_by(id)%>%
  mutate(oevarn_it = ifelse(is.na(oevar_it), mean(oevar_it, na.rm=TRUE), oevar_it))
```

#2)로짓 

```{r}
#상관관계 0.1 이하 변수들의 선택 
cor(datbyt[,c(3,4,6,7,8)])
```

```{r}
library(mlogit)
glm <- glm(n~ oe_it+ oesquare_it+levup_it, data = datbyt, family='binomial')
summary(glm)
```

```{r}
summary(glm)$coef[,4]
```
```{r}
glm.probs2 = predict(glm, type = 'response')
datbyt$probs2<- glm.probs2
datbyt<- datbyt%>%
  group_by(id)%>%
  mutate(pmean = mean(probs), 
         class = ifelse( mean(probs)>0.5, "A", "E"),
         pmean2 = mean(probs2), 
         class2 = ifelse( mean(probs2)>0.5, "A", "E"))

datclass <- datbyt[,c(2,12,14)]
datclass <- unique(datclass)

datclass%>%
  group_by(class2)%>%
  summarise(n=n())
```


3) kmeans 
```{r}
datbyt <- datbyt[,c(1,2,3,4,8,6,7)]

#scale
datbyt.scaled <-scale(datbyt[,3:6])
datbyt.scaled <-as.data.frame(datbyt.scaled)
datbyt$oe_itS <-datbyt.scaled$oe_it
datbyt$oesquare_itS <-datbyt.scaled$oesquare_it
datbyt$oevar_itS <-datbyt.scaled$oevarn_it
datbyt$levup_itS <-datbyt.scaled$levup_it

datbyt.id <-datbyt%>%
  group_by(id)%>%
  mutate(oe_iS = mean(oe_itS), 
         oesquare_iS = mean(oesquare_itS),
         oevar_iS = mean(oevar_itS),
         levup_iS = mean(levup_itS))

datbyt.id <- unique(datbyt.id[,c(2,13,14,15,16)])

```
```{r}
library(NbClust)
nc <- NbClust(datbyt[,8:11], min.nc=2, max.nc=6, method = 'kmeans')
#best는 4, 그 다음이 2 

nc2 <- NbClust(datbyt.id[,2:5], min.nc=2, max.nc=6, method = 'kmeans')
#best는 4, 그 다음이 2 
```

```{r}
kmeans.dat <- kmeans(datbyt.id[,2:5],center=2)
kmeans.dat

# Cluster means:
#        oe_iS oesquare_iS   oevar_iS  levup_iS
# 1 -0.1663305  -0.1682871 -0.1254368 0.0415205 #Experiencer
# 2  1.4049717   1.3695449  1.1794375 0.6317970 #Achiever

datbyt.id$leave <- ifelse(kmeans.dat$cluster==1,1,0)#1:experiencer 0: achiever

#경향성에 따른 최종 분류 
datbyt.id %>%
  group_by(leave)%>%
  summarise(n=n())

#g   n 
#0	254 : achiever	
#1	3022 : experiencer
```

```{r}
boxplot(oe_iS+oesquare_iS+oevar_iS+levup_iS~leave, data =datbyt.id)
```
```{r}
write.csv(datbyt.id, "34_kmeansbyt.csv")
```

#1) kmeans by레벨 20이하 (초반)의 정보
# - IV: t=1~4까지의 개인의 operational efficiency 평균(oe_it), operational efficiency 편차의 평균(oesqaure_i),개인의 누적된 oe의 변동 중 최신값 
#g   n 
#1	2613 : Experiencer 
#0	621 : Achiever 


#2) kmeans by 4주간의 정보(만렙 이전)  
# - IV: t=1~4까지의 주별 개인의 operational efficiency 평균(oe_it), operational efficiency 편차의 평균(oesqaure_it), operational efficiency의 분산(oevar_it), 레벨변화(levelup_it) 
# - DV : t기에 접속/미접속 여부 
#g   n 
#0	254 : achiever	
#1	3022 : experiencer
